WIP: Clustered trxs & random waypoint benchmarks

+++ sanity_check.cpp
@@ -48,7 +48,7 @@ int main(int argc, char** argv) {
     // x-location:  0.....1.....2......3......4
     //            t[0]  t[1]  t[2]   t[3]   t[4]
     // each transceiver is '1' unit away from the others
-    for(int i = 0; i < NUM_TRXS; ++i) {
+    for(size_t i = 0; i < NUM_TRXS; ++i) {
         auto& t = trxs[i];
         // setting parameters for t
         t.set_x(i + 0.01);
@@ -87,7 +87,7 @@ int main(int argc, char** argv) {
         }
         assert(sender.recv(&raw_msg, 0) == 0);
     } else { // another rank make sure messages get received
-        for(int i = 0; i < NUM_TRXS; ++i) {
+        for(size_t i = 0; i < NUM_TRXS; ++i) {
             auto& t = trxs[i];
             char* raw_msg;
             // receive data with a certain timeout
